Rename op_qubits to qubits_with_op. Move _to_tuple out of util. Remove unnecessary attributes.
