+++ qiskit/util.py
@@ -95,18 +95,3 @@ def _has_connection(hostname, port):
         return True
     except Exception:  # pylint: disable=broad-except
         return False
-
-
-def _to_tuple(values):
-    """
-    Return the input, sorted, and as a tuple.
-
-    Args:
-        values (Union[int, Iterable[int]]): An integer, or iterable of integers.
-    Returns:
-        tuple: The input values as a sorted tuple.
-    """
-    try:
-        return tuple(values)
-    except TypeError:
-        return (values,)
